+++ kosogor/src/main/kotlin/tanvd/kosogor/defaults/Wrapper.kt
@@ -3,12 +3,12 @@ package tanvd.kosogor.defaults
 import org.gradle.api.Project
 import org.gradle.api.tasks.wrapper.Wrapper
 
-data class WrapperConfig(var version: String = "5.1.1", var distributionUrl: String? = null)
+data class WrapperConfig(var version: String = "5.2.1", var distributionUrl: String? = null)
 
 /**
  * Configure wrapper task of this project. Should be used only on root project.
  *
- * By default uses version 5.1.1
+ * By default uses version 5.2.1
  *
  * Used in default chain of Kosogor plugin.
  */