+++ kosogor/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+# 1.0.2 - 2019-03-13
+- Fix BuildDir default broke in 1.0.2 (it was not applied to parent projects)
+- Update wrapper default to 5.2.1
+- Add `userOrg` field for bintray plugin
+
 # 1.0.1 - 2019-03-7
 ### Added
 - PublishPlugin with support of Gradle plugins portal